requirements.txt
openai
qdrant-client
pyyaml
chardet
thefuzz
python-Levenshtein
tqdm
